First commit - create react app set up, plus prototype core function
This commit opens the repository. I used create react app to set up this project and installed a number of dependencies to allow us to use SCSS.

I worked on defining a basic app structure to be used as a spring board, and in doing so implemented a first pass of the core function of our app. This function is the TalkBoard which allows users to highlight images and create stories from those images.

Also installed ESLINT for helping us to work on code in a similar fashion across the team.