Added api for getting images for gallery along with modal and new form
This commit contains quite a few bits. Including a lot of refactoring to take the new api calls out of the components and into module calls or HOC

I have also added a modal for the gallery which will allow users to pop in a category name or the images they want to display.