Refactor completely reworked redux to follow better guidelines
We now intiialize the state of user gallery and categoires in the image agallery container componenent. Added redux to allow the asynchronous fetching of this data. Then reworked redux actions into an action file and use dispatch as it is shown in the docs by calling a function which returns the object we need.

Also changed up our reducer function to user object assign, this is also similar to the docs and is less verbose.